requirements.txt
arxiv>=2.2.0
chromadb>=1.0.20
langchain>=0.3.27
langchain-community>=0.3.29
langchain-core>=0.3.76
langchain-openai>=0.3.33
langchain-chroma>=0.2.5
sentence-transformers>=5.1.0
openai>=1.107.1
pydantic>=2.11.7
python-dotenv>=1.1.1
numpy>=2.3.3
scikit-learn>=1.7.2
scipy>=1.16.1
torch>=2.8.0
transformers>=4.56.1
tiktoken>=0.11.0
PyMuPDF>=1.26.4
rich>=14.1.0
typer>=0.17.4
requests>=2.32.5
